requirement.txt
streamlit==1.24.1
faiss-cpu==1.7.3
numpy==1.26.4
sentence-transformers==2.2.2
torch==2.0.1
google-generativeai==0.2.1



